requirements.txt
Django==4.0.1
Pillow==9.0.0
django-tinymce==3.4.0
